requirements.txt
absl-py==0.1.9
astroid==1.6.1
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
futures==3.2.0
html5lib==0.9999999
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
numpy==1.14.0
Pillow==5.0.0
pkg-resources==0.0.0
protobuf==3.5.1
pylint==1.8.2
python-resize-image==1.1.11
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
